chore: move events to new module
